Introduce crypto::common::DigestContext, move nsec3_hash to dnssec::common.
