+++ src/dnssec/sign/test_util/mod.rs
@@ -10,7 +10,7 @@ use bytes::Bytes;
 use crate::base::iana::{Class, SecAlg};
 use crate::base::name::FlattenInto;
 use crate::base::{Name, Record, Rtype, Serial, ToName, Ttl};
-use crate::crypto::validate::nsec3_hash;
+use crate::dnssec::common::nsec3_hash;
 use crate::dnssec::sign::denial::nsec3::mk_hashed_nsec3_owner_name;
 use crate::rdata::dnssec::{RtypeBitmap, Timestamp};
 use crate::rdata::nsec3::OwnerHash;
