+++ src/dnssec/validator/nsec.rs
@@ -12,8 +12,8 @@ use crate::base::iana::{ExtendedErrorCode, Nsec3HashAlg};
 use crate::base::name::{Label, ToName};
 use crate::base::opt::ExtendedError;
 use crate::base::{Name, ParsedName, Rtype};
-use crate::crypto::validate::nsec3_hash;
 use crate::dep::octseq::Octets;
+use crate::dnssec::common::nsec3_hash;
 use crate::rdata::nsec3::{Nsec3Salt, OwnerHash};
 use crate::rdata::{AllRecordData, Nsec, Nsec3};
